consolidate implementations of IR function regisration
